requirements.txt
aiohttp==3.10.5
botbuilder-core>=4.14.0
botbuilder-dialogs>=4.14.0
botbuilder-integration-aiohttp>=4.14.0
botframework-connector>=4.14.0
python-dotenv>=1.0.0
openai>=1.50.0
azure-ai-vision-imageanalysis==1.0.0
azure-ai-language-conversations==1.1.0
azure-search-documents==11.6.0
azure-identity>=1.15.0
azure-core>=1.30.0
azure-cognitiveservices-speech>=1.34.0